WIP MusicDisplayer Pause Resume
WIP changing MusicDisplayer behaviour
WIP added pausing and resuming abilities
WIP added message safe deletion so the the bot can only delete its own message even if its rights allow it to bypass, in prevision to the change to `SlashCommands`

+++ modules/botModules/MessagePrintReply.mjs
@@ -28,7 +28,8 @@ const errorIcon = `https://cdn.discordapp.com/attachments/329613279204999170/970
                 setTimeout(
                     () => {
                         if (msg)
-                        msg.delete().catch(()=>{});
+                            if (msg.author.bot)
+                                msg.delete().catch(()=>{});
                     },
                     time * 1000);
             }
